requirements.txt
altgraph==0.17.4
filelock==3.18.0
fsspec==2025.3.0
importlib_metadata==8.6.1
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pywin32-ctypes==0.2.3
sympy==1.13.1
torch @https://download.pytorch.org/whl/cu124/torch-2.5.1%2Bcu124-cp39-cp39-win_amd64.whl#sha256=9036c4372dec409842a80965d94b7b0fb4298e0967ceb03336a42c83778faa6f
torchaudio==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
zipp==3.21.0
